requirements.txt
# Unsloth for efficient fine-tuning
unsloth[colab-new] @ git+https://github.com/unslothai/unsloth.git
unsloth_zoo @ git+https://github.com/unslothai/unsloth-zoo.git

# Core ML
torch>=2.1.0
transformers>=4.45.0
datasets>=2.16.0
accelerate>=0.25.0
peft>=0.7.0
bitsandbytes>=0.41.0
trl>=0.7.0

# Logging and tracking
wandb>=0.16.0

# HuggingFace Hub
huggingface_hub>=0.20.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
numpy>=1.24.0

torchvision